Revert "split single collector into series of collectors"
This reverts commit c2c5d5d848481fb874b65f48917c0769808c317b.
